moved collection_metadata to bhom/wrapped, and added method that gets montly diurnal ranges
